Fix re-nomination problem
In the current design, we do not use a timer to trigger nominations.
Instead, a new transaction being received can trigger the nomination
protocol if there are enough transactions in the pool to form a
new block.

There is a timing issue where a new transaction could be received
while there was already a nomination round in progress.

Fix the issue by temporarily adding a status flag, signaling
whether we are currently in a nomination round.

This will later be replaced with the use of timers (real / virtual),
ensuring that a nomination is started in predictable time periods.