Do not nominate blocks if we're a full node
Only Validator nodes are allowed to create new blocks.

Added a sleep call between creation of blocks in a loop for one test-case,
which seems to fail when the entire agor atest-suite is running.
The assumption is that this is some kind of timing issue with leftover
SCP messages being still in flight, but it needs more investigation.

todo: investigate more.