Fix the NetworkManager tests
The test needs to consist of one validator,
one full node (victim), and another full node (byzantine).

The Validator is the only node allowed to create blocks,
while the victim node tries to retrieve blocks via getBlocksFrom().

The victim will reject blocks from the byzantine node because
the blocks are not cryptographically valid.