Fix the timer handling in setupTimer()
When SCP calls setupTimer() with a timeout of zero,
it wants to cancel all previously started timers.

However, there was a timing issue. Previously we added and
removed timers based on the 'timerIDs', which are not actually
IDs but are timer types (there are only two types: ballot and nominator).

A possible scenario was as follows (pseudocode):

- setupTimer(1.seconds, ID.Ballot, { firstVote(); });
- ... 500 msecs later
- setupTimer(0.seconds, ID.Ballot, null);
  => this "deactivates the timer"
- ... 100 msecs later
- setupTimer(1.seconds, ID.Ballot, { secondVote(); });
- ... 400 msecs later
  => the first timer now gets invoked, even though it was "cancelled"

The reason for this bug is because we used a single hashmap for all timers
based on the timer type.

This has been changed, so now all timers have a unique incrementing ID,
and when a timer of a specific type is cancelled, then all *current* timers
will be cancelled, but any new ones will continue to work.