Sign and verify SCPStatements
This uses the node's public key to sign SCPStatements
it wants to propagate in SCPEnvelopes, and also uses
the key for validation of the signature.

TODO: Should use envelope.statement.nodeID!

TODO: It seems Herder in SCP signs both envelopes (SCP),
and also StellarValue, which is something we want.

TODO: So LedgerCloseValueSignature() contains the node which
introduced the StellarValue. This could mean that in our election
process, we have one node which collects signatures and then
publishes the Schnorr sig of that block.

Then, each SCP message is also additionally signed too.

TODO: check lcValueSignature() and validateValueHelper()

TODO: check when validateValue() is actually called, if it's
at the end, then it should be ok.

TODO: Check the order of stellar nominator routines.