Implement a Quorum config generator algorithm
The generator uses a provided source of randomness
to shuffle the nodes into different quorums.

The source of randomness should be derived from
the revealed preimages of the validators.

Nodes which have a higher staken amount have a
higher chance of being included in each node's
quorum set.

Part of: #240