Split Node into FullNode and Validator
A FullNode should not contain any code related
to running the consensus protocol.

The separation of concerns avoid proliferation of
'if' statements in the Node class, and moves
the responsibility of the validator explicitly
into the Validator class.

In the unittests we instantiate RemoteAPI with
agora.API.Validator to avoid having to introduce
a lot of templated code and special cases.

When the NetworkManager tries to retrieve the PublicKey
of a node (which is actually a FullNode), the Node
in the unittests will return PublicKey.init.

In production, vibe.d will actually produce an
HTTPStatusException with status == 404,
so we handle it with a try/catch.

Also changed the system integration tests to
instantiate a FullNode (in addition to the existing
3 Validator nodes), to ensure the 404 HTTP status
handing code is actually executed.

Co-Authored-By: Andrej Mitrovic <andrej.mitrovich@gmail.com>