Implement ScopeCondition for conditional support
It's used for tracking which conditional scopes are true where
the code may be executed, or not true where code execution will be
skipped until we exit into a scope which is true.

Largely based on Bitcoin's ConditionStack implementation,
with simplifications and a series of tests added.

MIT license attribution was added.