Flash: Implement Invoice definition
The createInvoice() routine generates the invoice
and the associated secret.

The hash of the secret is stored
in the invoice together with the invoice's amount, expiry,
and any description.

The secret should be shared with the party which will pay the invoice.
The hash of the secret is used in the contract which is shared across
zero or more channel hops. The payment is realized once the payer reveals
the secret, and all the channel pairs in the multi-hop channel layout
have their settle/update transactions signed.